Fix: validate empty string

+++ app/src/main/java/edu/arimanius/letsqueeze/ui/setting/SettingFragment.kt
@@ -104,7 +104,7 @@ class SettingFragment : Fragment() {
 
             override fun afterTextChanged(s: Editable) {
                 settingViewModel.numQuestionChanged(
-                    numQuestionEditText.text.toString().toInt(),
+                    numQuestionEditText.text.toString(),
                 )
             }
         })
